Should not throw an error if the temp dir does not yet exist.
git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@1222711 13f79535-47bb-0310-9956-ffa450edef68

+++ src/tests/junit/org/apache/tools/ant/taskdefs/optional/junit/XMLResultAggregatorTest.java
@@ -41,7 +41,9 @@ public class XMLResultAggregatorTest extends TestCase {
             return;
         }
         final File d = new File(System.getProperty("java.io.tmpdir"), "XMLResultAggregatorTest");
-        new Delete() {{removeDir(d);}}; // is there no utility method for this?
+        if (d.exists()) {
+            new Delete() {{removeDir(d);}}; // is there no utility method for this?
+        }
         assertTrue(d.getAbsolutePath(), d.mkdir());
         File xml = new File(d, "x.xml");
         PrintWriter pw = new PrintWriter(new FileOutputStream(xml));
